Revert to int64

+++ coderd/insights.go
@@ -207,7 +207,7 @@ func (api *API) insightsTemplates(rw http.ResponseWriter, r *http.Request) {
 				StartTime:   startTime,
 				EndTime:     endTime,
 				TemplateIDs: templateIDs,
-				Interval:    24 * time.Hour,
+				Interval:    int64(24 * time.Hour),
 			})
 			if err != nil {
 				return xerrors.Errorf("get template daily insights: %w", err)
